+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/viewmodel/VoucherDialogViewModel.kt
@@ -1,0 +1,84 @@
+package net.mullvad.mullvadvpn.viewmodel
+
+import android.content.res.Resources
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.launch
+import net.mullvad.mullvadvpn.R
+import net.mullvad.mullvadvpn.compose.state.LoginUiState
+import net.mullvad.mullvadvpn.compose.state.VoucherDialogState
+import net.mullvad.mullvadvpn.compose.state.VoucherDialogUiState
+import net.mullvad.mullvadvpn.model.VoucherSubmissionError
+import net.mullvad.mullvadvpn.model.VoucherSubmissionResult
+import net.mullvad.mullvadvpn.ui.serviceconnection.ServiceConnectionManager
+import net.mullvad.mullvadvpn.ui.serviceconnection.VoucherRedeemer
+import org.joda.time.DateTimeConstants.SECONDS_PER_DAY
+
+class VoucherDialogViewModel(
+    serviceConnectionManager: ServiceConnectionManager,
+    private val resources: Resources
+) : ViewModel() {
+
+    private val vmState = MutableStateFlow<VoucherDialogState>(VoucherDialogState.Default)
+    private val voucherInput = MutableStateFlow(LoginUiState.INITIAL.accountNumberInput)
+
+    private var voucherRedeemer: VoucherRedeemer? =
+        serviceConnectionManager.connectionState.value.readyContainer()?.voucherRedeemer
+
+    var uiState =
+        combine(vmState, voucherInput) { state, input ->
+                VoucherDialogUiState(voucherInput = input, voucherViewModelState = state)
+            }
+            .stateIn(viewModelScope, SharingStarted.WhileSubscribed(), VoucherDialogUiState.INITIAL)
+
+    fun onRedeem(voucherCode: String) {
+        vmState.update { VoucherDialogState.Verifying }
+        viewModelScope.launch {
+            when (val result = voucherRedeemer?.submit(voucherCode)) {
+                is VoucherSubmissionResult.Ok -> handleAddedTime(result.submission.timeAdded)
+                is VoucherSubmissionResult.Error -> setError(result.error)
+                else -> {
+                    vmState.update { VoucherDialogState.Default }
+                }
+            }
+        }
+    }
+
+    fun onVoucherInputChange(voucherString: String) {
+        voucherInput.value = voucherString
+    }
+
+    private fun handleAddedTime(timeAdded: Long) {
+        val days = (timeAdded / SECONDS_PER_DAY).toInt()
+        viewModelScope.launch {
+            vmState.update {
+                VoucherDialogState.Success(
+                    if (days > 60) {
+                        resources.getString(R.string.months_added_to_your_account, days / 60)
+                    } else {
+                        resources.getQuantityString(R.plurals.days_added_to_your_account, days)
+                    }
+                )
+            }
+        }
+    }
+
+    private fun setError(error: VoucherSubmissionError) {
+        viewModelScope.launch {
+            val message =
+                resources.getString(
+                    when (error) {
+                        VoucherSubmissionError.InvalidVoucher -> R.string.invalid_voucher
+                        VoucherSubmissionError.VoucherAlreadyUsed -> R.string.voucher_already_used
+                        else -> R.string.error_occurred
+                    }
+                )
+            vmState.update { VoucherDialogState.Error(message) }
+        }
+    }
+}