+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Line wrap the file at 100 chars.                                              Th
 - Move out of time evaluation to connect view model.
 - Migrate out of time view to compose.
 - Migrate login view to compose.
+- Migrate voucher dialog to compose.
 
 #### Linux
 - Don't block forwarding of traffic when the split tunnel mark (ct mark) is set.